Added missing dependencies
* dependencies for `double-conversion`and `gettext`have been added
* sha256 has been changed

after that, the compilation works successfully, however, the installation fails because `/usr/local/share/inkscape` was not created. The problem seems to be related to the installation directory and the fact that is protected by MacOS' SIP. I tried changing `LIBPREFIX`to `ENV.prepend_path "LIBPREFIX", prefix` (resulting in the variable set to `/usr/local/Cellar/inkscape/master`, which I think is correct), but I think there is code that needs to be fixed in Inkscapes build files. 

So in essence: Build now works, but the installation fails.